pyproject.toml
[tool.poetry]
name = "Alzheimer AlexNet"
version = "1.0"
description = ""
authors = ["Adél Bajcsi <adel.bajcsi@ubbcluj.ro>"]
readme = "README.md"
repository = "https://github.com/bajcsiadel/Reproducing-Predicting-Alzheimer"
packages = [
	{ include = "model", from = "src" },
	{ include = "utils", from = "src" },
]
include = [
	"README.md",
	"src/conf",
	"src/preprocessing",
	"src/scripts",
]

[[tool.poetry.source]]
name = "torch+cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "supplemental"

[tool.poetry.dependencies]
python = "^3.11,<3.13"
albumentations = "^1.4.10"
hydra-core = "^1.3.2"
icecream = "^2.1.3"
# numpy is required for albumentations, matplotlib, seaborn, and torch
# matplotlib is required for seaborn
nilearn = "^0.10.4"
opencv-contrib-python = "^4.8.1"
# pandas is required for seaborn
pipe = "^2.0"
pydicom = "^2.4.4"
python-dotenv = "^1.0.0"
seaborn = "^0.13.2"
# scikit-learn is reuqired for albumentations
tensorboard = "^2.16.2"
torch = [
	{version = "^2.3.0", source = "pypi", markers = "sys_platform == 'darwin'"},
	{version = "2.3.0+cu121", source = "torch+cu121", markers = "sys_platform != 'darwin'"}
]
torchinfo = "^1.8.0"
torchvision = [
	{version = "^0.18.0", source = "pypi", markers = "sys_platform == 'darwin'"},
	{version = "0.18.0+cu121", source = "torch+cu121", markers = "sys_platform != 'darwin'"}
]
tqdm = "^4.66.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
flake8-bugbear = "^23.12.2"
isort = "^5.12.0"
pre-commit = "^3.4.0"
pyproject-flake8 = "^6.1.0"

[tool.black]
line-length = 88

[tool.flake8]
ignore = ["E203", "E302", "E501", "E999", "F541", "W503"]
max-line-length = 88
max-complexity = 12
select = ["B", "C", "E", "F", "W", "B9"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
